Enhancing the gitconfig
This commit changes the following settings.

Editor:
-------

The editor is now less -FRSX wish respect color codes and use a paging
mode only if the command output is longer than the screen. Pretty useful
for diff, log or lg commands.

Status:
-------

Adding -sb to git status. This will show the status in short version
which has proven to be more readable for students. It also displays the
distance (in commits) to the remote branch.

Pretty log:
-----------

The log format is also cleaned up since --abbrev-commit and --date=relative
are already provided by %h and %cr respectively.

The branches list has been move to the end of the line to keep consistency
between lines.

Finally, the --all flag shows every branches and stashed commits.

Autocorrect:
------------

The help.autocorrect option tells git not to troll you by saying
'Did you mean git log' when you type 'git lgo' and other little mistakes.

Unstage:
--------

The git unstage alias was added to allow unstaging files that have been
added to the index. This is useful to "unadd" a file or unmark a
conflicting file.